+++ Mindspace_front_next/app/map/map.tsx
@@ -3,7 +3,7 @@ import { useEffect, useRef, useState } from "react";
 import { ForceGraph2D } from "react-force-graph";
 import { NodeObject, Node, Context } from "@/constants/types";
 import { useNodeListQuery } from "@/api/hooks/queries/node";
-import Loading from "@/components/Loading";
+import Loading from "@/components/MoonLoadSpinner";
 import NodeModal from "./components/Modal";
 
 import { useRecoilState } from "recoil";
